It worked? Change the way we inspect sockets
Previous version would check every potential file descriptor, not just
the one corresponding to sockets (And hence those part of the select).
And since we used the fd as array index it would cause a buffer
overflow if the size of the array of sockets was larger smaller than
the maximum file descriptor (and it usually is the case, since they are
at least the usual std{err,in,out}).
The glibc would cry and valgrind would also complain.